🦄 refactor(pages): move Header & Sidebar into layout
